fix: merge conv2d split files back into test_conv2d.py, use l1_small_size=32768
Module-scoped device causes L1_SMALL allocations to accumulate across
the 192 test_conv_features parametrized cases. With per-function device
each test got a fresh 16384-byte pool; with module device previous tests
leave ~16KB allocated by the time the large 256x256 HEIGHT_SHARDED config
runs, causing OOM (224 bytes free, needs 544 per bank).

Fix: consolidate test_conv2d_l1_16384.py + test_conv2d_l1_32768.py back
into test_conv2d.py with l1_small_size=32768. test_conv_dram already
required 32768; bumping test_conv_features to the same value gives
sufficient headroom across the full module run.
